Refactor XML generator
