+++ examples/src/main/java/edu/berkeley/cs/jqf/examples/xml/XMLDocumentUtils.java
@@ -1,0 +1,71 @@
+/*
+ * Copyright (c) 2019, The Regents of the University of California
+ *
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are
+ * met:
+ *
+ * 1. Redistributions of source code must retain the above copyright
+ * notice, this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright
+ * notice, this list of conditions and the following disclaimer in the
+ * documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package edu.berkeley.cs.jqf.examples.xml;
+
+import javax.xml.transform.Transformer;
+import javax.xml.transform.TransformerException;
+import javax.xml.transform.TransformerFactory;
+import javax.xml.transform.dom.DOMSource;
+import javax.xml.transform.stream.StreamResult;
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.InputStream;
+import java.io.StringWriter;
+
+import org.w3c.dom.Document;
+
+/**
+ * @author Rohan Padhye
+ */
+public class XMLDocumentUtils {
+    private static TransformerFactory transformerFactory =
+            TransformerFactory.newInstance();
+
+    public static String documentToString(Document document) {
+        try {
+            Transformer transformer = transformerFactory.newTransformer();
+            StringWriter stream = new StringWriter();
+            transformer.transform(new DOMSource(document), new StreamResult(stream));
+            return stream.toString();
+        } catch (TransformerException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public static InputStream documentToInputStream(Document document) {
+        try {
+            Transformer transformer = transformerFactory.newTransformer();
+            ByteArrayOutputStream stream = new ByteArrayOutputStream();
+            transformer.transform(new DOMSource(document), new StreamResult(stream));
+            return new ByteArrayInputStream(stream.toByteArray());
+        } catch (TransformerException e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
