+++ examples/src/test/java/edu/berkeley/cs/jqf/examples/maven/ModelReaderTest.java
@@ -33,7 +33,8 @@ import java.io.IOException;
 import java.io.InputStream;
 
 import com.pholser.junit.quickcheck.From;
-import edu.berkeley.cs.jqf.examples.XmlDocumentGenerator;
+import edu.berkeley.cs.jqf.examples.xml.XMLDocumentUtils;
+import edu.berkeley.cs.jqf.examples.xml.XmlDocumentGenerator;
 import edu.berkeley.cs.jqf.examples.common.Dictionary;
 import edu.berkeley.cs.jqf.fuzz.Fuzz;
 import edu.berkeley.cs.jqf.fuzz.JQF;
@@ -63,13 +64,13 @@ public class ModelReaderTest {
     @Fuzz
     public void testWithGenerator(@From(XmlDocumentGenerator.class)
                                       @Dictionary("dictionaries/maven-model.dict") Document dom) {
-        testWithInputStream(XmlDocumentGenerator.documentToInputStream(dom));
+        testWithInputStream(XMLDocumentUtils.documentToInputStream(dom));
     }
 
     @Fuzz
     public void debugWithGenerator(@From(XmlDocumentGenerator.class)
                                        @Dictionary("dictionaries/maven-model.dict") Document dom) {
-        System.out.println(XmlDocumentGenerator.documentToString(dom));
+        System.out.println(XMLDocumentUtils.documentToString(dom));
         testWithGenerator(dom);
     }
