+++ examples/src/test/java/edu/berkeley/cs/jqf/examples/tomcat/WebXmlTest.java
@@ -32,7 +32,8 @@ import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 
 import com.pholser.junit.quickcheck.From;
-import edu.berkeley.cs.jqf.examples.XmlDocumentGenerator;
+import edu.berkeley.cs.jqf.examples.xml.XMLDocumentUtils;
+import edu.berkeley.cs.jqf.examples.xml.XmlDocumentGenerator;
 import edu.berkeley.cs.jqf.examples.common.Dictionary;
 import edu.berkeley.cs.jqf.fuzz.Fuzz;
 import edu.berkeley.cs.jqf.fuzz.JQF;
@@ -58,12 +59,12 @@ public class WebXmlTest {
 
     @Fuzz
     public void testWithGenerator(@From(XmlDocumentGenerator.class) @Dictionary("dictionaries/tomcat-webxml.dict") Document dom) {
-        testWithInputStream(XmlDocumentGenerator.documentToInputStream(dom));
+        testWithInputStream(XMLDocumentUtils.documentToInputStream(dom));
     }
 
     @Fuzz
     public void debugWithGenerator(@From(XmlDocumentGenerator.class) @Dictionary("dictionaries/tomcat-webxml.dict") Document dom) {
-        System.out.println(XmlDocumentGenerator.documentToString(dom));
+        System.out.println(XMLDocumentUtils.documentToString(dom));
         testWithGenerator(dom);
     }
